[4.x.x.x] Change product rating field type to float

+++ upload/admin/model/catalog/review.php
@@ -161,7 +161,7 @@ class Review extends \Opencart\System\Engine\Model {
 	 *
 	 * @param int $product_id primary key of the product record
 	 *
-	 * @return int total number of rating records that have product ID
+	 * @return float average rating that have product ID
 	 *
 	 * @example
 	 *
@@ -169,11 +169,11 @@ class Review extends \Opencart\System\Engine\Model {
 	 *
 	 * $rating_info = $this->model_catalog_review->getRating($product_id);
 	 */
-	public function getRating(int $product_id): int {
+	public function getRating(int $product_id): float {
 		$query = $this->db->query("SELECT AVG(`rating`) AS `total` FROM `" . DB_PREFIX . "review` WHERE `product_id` = '" . (int)$product_id . "' AND `status` = '1'");
 
 		if ($query->num_rows) {
-			return (int)$query->row['total'];
+			return (float)$query->row['total'];
 		} else {
 			return 0;
 		}
