+++ upload/catalog/model/catalog/product.php
@@ -72,7 +72,7 @@ class Product extends \Opencart\System\Engine\Model {
 			$product_data['variant'] = $query->row['variant'] ? json_decode($query->row['variant'], true) : [];
 			$product_data['override'] = $query->row['override'] ? json_decode($query->row['override'], true) : [];
 			$product_data['price'] = (float)($query->row['discount'] ?: $query->row['price']);
-			$product_data['rating'] = (int)$query->row['rating'];
+			$product_data['rating'] = (float)$query->row['rating'];
 			$product_data['reviews'] = (int)$query->row['reviews'] ? $query->row['reviews'] : 0;
 
 			return $product_data;
